app/requirements.txt
# requirements.txt
confluent-kafka==2.3.0
python-dotenv==1.0.0